scaffold diplan geometry interactions
