+++ packages/clients/diplan/src/store/geoEditing/duplicatePolygons.ts
@@ -1,0 +1,8 @@
+import { PolarActionContext } from '@polar/lib-custom-types'
+import { DiplanGetters, DiplanState } from '../../types'
+
+export const duplicatePolygons = ({
+  dispatch,
+}: PolarActionContext<DiplanState, DiplanGetters>) => {
+  dispatch('plugin/draw/setMode', 'none', { root: true })
+}
